Исправляет неточности после ревью
