+++ js/form.js
@@ -12,38 +12,6 @@ const closeImgEditingForm = document.querySelector('#upload-cancel');
 const textHashtagsField = document.querySelector('.text__hashtags');
 const textDescriptionField = document.querySelector('.text__description');
 
-const successTemplate = document.querySelector('#success').content.querySelector('.success');
-const successElement = successTemplate.cloneNode(true);
-const successButton = successElement.querySelector('.success__button');
-
-const errorTemplate = document.querySelector('#error').content.querySelector('.error');
-const errorElement = errorTemplate.cloneNode(true);
-const errorButton = errorElement.querySelector('.error__button');
-
-const clearForm = () => {
-  uploadFileElement.value = '';
-  textHashtagsField.value = '';
-  textDescriptionField.value = '';
-
-  // imgUploadForm.reset();
-};
-
-const onDocumentKeydown = (evt) => {
-  if (isEscapeKey(evt)) {
-    evt.preventDefault();
-    imgUploadElement.classList.add('hidden');
-    bodyElement.classList.remove('modal-open');
-  }
-};
-
-textHashtagsField.addEventListener('keydown', (evt) => {
-  evt.stopPropagation();
-});
-
-textDescriptionField.addEventListener('keydown', (evt) => {
-  evt.stopPropagation();
-});
-
 const pristine = new Pristine(imgUploadForm, {
   classTo: 'img-upload__field-wrapper',
   errorClass: 'img-upload__field-wrapper--invalid',
@@ -59,9 +27,9 @@ pristine.addValidator(textDescriptionField,
   'Максимальная длина 140 символов'
 );
 
-const checkLangthHashtag = (value) => value.length <= MAX_HASHTAG_LENGTH;
+const checkLengthHashtag = (value) => value.length <= MAX_HASHTAG_LENGTH;
 
-const checkSimbolsHashtag = (item) => HASHTAG.test(item);
+const checkSymbolsHashtag = (item) => HASHTAG.test(item);
 
 const checkRepeatHashtag = (item) => {
   const lowerHashtag = item.map((tag) => tag.toLowerCase());
@@ -69,8 +37,8 @@ const checkRepeatHashtag = (item) => {
 };
 
 const validateHashtag = (value) => {
-  const splitHashtag = value.trim().split(' ');
-  return checkLangthHashtag(splitHashtag) && splitHashtag.every(checkSimbolsHashtag) && checkRepeatHashtag(splitHashtag);
+  const hashtags = value.trim().split(' ').filter((tag) => tag.trim().length);
+  return checkLengthHashtag(hashtags) && hashtags.every(checkSymbolsHashtag) && checkRepeatHashtag(hashtags);
 };
 
 pristine.addValidator(textHashtagsField,
@@ -78,70 +46,48 @@ pristine.addValidator(textHashtagsField,
   'Хэштег введен неверно'
 );
 
-const onErrorKeydown = (evt) => {
-  // const errorSection = document.querySelector('.error');
-  if (isEscapeKey(evt)) {
+const clearForm = () => {
+  imgUploadForm.reset();
+  pristine.reset();
+};
+
+const closeFormModal = () => {
+  clearForm();
+  imgUploadElement.classList.add('hidden');
+  bodyElement.classList.remove('modal-open');
+  textHashtagsField.addEventListener('keydown', (evt) => {
     evt.stopPropagation();
-    errorElement.classList.add('hidden');
+  });
+  textDescriptionField.addEventListener('keydown', (evt) => {
+    evt.stopPropagation();
+  });
+};
+
+const openFormModal = () => {
+  imgUploadElement.classList.remove('hidden');
+  bodyElement.classList.add('modal-open');
+};
+
+const onDocumentKeydown = (evt) => {
+  if (isEscapeKey(evt)) {
+    evt.preventDefault();
+    closeFormModal();
   }
 };
 
-const openErrorMessage = () => {
-  bodyElement.appendChild(errorElement);
-  errorButton.addEventListener('click', () => {
-    errorElement.classList.add('hidden');
-  });
-
-  // errorElement.addEventListener('keydown', (evt) => {
-  //   if (isEscapeKey(evt)) {
-  //     evt.stopPropagation();
-  //   }
-  // });
-
-  // const errorSection = document.querySelector('.error-inner');
-  // errorSection.addEventListener('keydown', onErrorKeydown);
-  errorElement.addEventListener('keydown', onErrorKeydown);
-};
-
-const openSuccessMessage = () => {
-  bodyElement.appendChild(successElement);
-  successButton.addEventListener('click', () => {
-    successElement.classList.add('hidden');
-    clearForm();
-    imgUploadElement.classList.add('hidden');
-  });
-
-  // successElement.addEventListener('keydown', (evt) => {
-  //   if (isEscapeKey(evt)) {
-  //     evt.stopPropagation();
-  //   }
-  // });
-};
-
 uploadFileElement.addEventListener('change', () => {
-  imgUploadElement.classList.remove('hidden');
-  bodyElement.classList.add('modal-open');
-  pristine.reset();
-  // clearForm();
-
+  openFormModal();
   document.addEventListener('keydown', onDocumentKeydown);
 });
 
 closeImgEditingForm.addEventListener('click', () => {
-  imgUploadElement.classList.add('hidden');
-  bodyElement.classList.remove('modal-open');
-
+  closeFormModal();
   document.removeEventListener('keydown', onDocumentKeydown);
 });
 
 imgUploadForm.addEventListener('submit', (evt) => {
-  evt.preventDefault();
-
   const isValid = pristine.validate();
   if (!isValid) {
-    openErrorMessage();
-  } else {
-    openSuccessMessage();
+    evt.preventDefault();
   }
-  // this.submit();
 });
